move fixtures into conftest.py
