+++ test/test_composite_solution.py
@@ -11,67 +11,15 @@ import pytest
 from solvis import CompositeSolution, FaultSystemSolution
 from solvis.inversion_solution.inversion_solution import BranchInversionSolution, InversionSolution
 
+from test.conftest import branch_solutions
 current_model = nm.get_model_version(nm.CURRENT_VERSION)
 slt = current_model.source_logic_tree()
 fslt = slt.fault_system_branches[0]  # PUY is used always , just for the 3 solution_ids
 
-ARCHIVES = dict(
-    CRU="ModularAlpineVernonInversionSolution.zip",
-    HIK="AveragedHikurangiInversionSolution-QXV0b21hdGlvblRhc2s6MTA3MzMy.zip",
-    PUY="PuysegurInversionSolution-QXV0b21hdGlvblRhc2s6MTExMDA1.zip",
-)
 
 FSR_COLUMNS_A = 26
 FSR_COLUMNS_B = 25  # HIK
 RATE_COLUMNS_A = 6
-
-
-def get_solution(id: str, archive: str) -> InversionSolution:
-    files = dict(
-        U2NhbGVkSW52ZXJzaW9uU29sdXRpb246MTE4NTQ2=archive,
-        U2NhbGVkSW52ZXJzaW9uU29sdXRpb246MTE4NTQz=archive,
-        U2NhbGVkSW52ZXJzaW9uU29sdXRpb246MTE4NTQ1=archive,
-    )
-    folder = pathlib.PurePath(os.path.realpath(__file__)).parent
-    filename = pathlib.PurePath(folder, f"fixtures/{files[id]}")
-    return InversionSolution().from_archive(str(filename))
-
-
-def branch_solutions(fslt, archive=ARCHIVES['CRU'], rupt_set_id='RUPTSET_ID'):
-    for branch in fslt.branches:
-        yield BranchInversionSolution.new_branch_solution(
-            get_solution(branch.inversion_solution_id, archive), branch, fslt.short_name, rupt_set_id
-        )
-
-
-@pytest.fixture(scope='class')
-def composite_fixture():
-
-    # solutions = []
-
-    v1_0_0 = nm.get_model_version('NSHM_1.0.0')
-    slt = v1_0_0.source_logic_tree()
-    print(slt.fault_system_branches[0])
-
-    # fudge the model branches because we have too few fixtures
-    for idx in [1, 2]:
-        slt.fault_system_branches[idx].branches = deepcopy(slt.fault_system_branches[0].branches)
-
-    composite = CompositeSolution(slt)  # create the new composite solutoin
-    for fault_system_lt in slt.fault_system_branches:
-        if fault_system_lt.short_name in ['CRU', 'PUY', 'HIK']:
-            composite.add_fault_system_solution(
-                fault_system_lt.short_name,
-                FaultSystemSolution.from_branch_solutions(
-                    branch_solutions(
-                        fault_system_lt,
-                        archive=ARCHIVES[fault_system_lt.short_name],
-                        rupt_set_id=f'rupset_{fault_system_lt.short_name}',
-                    )
-                ),
-            )
-    return composite
-
 
 class TestThreeFaultSystems(object):
 
@@ -86,6 +34,7 @@ class TestThreeFaultSystems(object):
     @pytest.mark.TODO_check_values
     def test_rupture_surface(self, composite_fixture):
         surface = composite_fixture.rupture_surface('PUY', 3)
+        print(surface)
         assert surface.shape == (5, FSR_COLUMNS_A)
 
     @pytest.mark.TODO_check_values
@@ -103,7 +52,7 @@ class TestThreeFaultSystems(object):
         assert surfaces.shape == (809, 15)
 
 
-def test_composite_serialisation():
+def test_composite_serialisation(archives):
     folder = tempfile.TemporaryDirectory()
     # folder = pathlib.PurePath(os.path.realpath(__file__)).parent
 
@@ -121,7 +70,7 @@ def test_composite_serialisation():
             solutions = list(
                 branch_solutions(
                     fault_system_lt,
-                    archive=ARCHIVES[fault_system_lt.short_name],
+                    archive=archives[fault_system_lt.short_name],
                     rupt_set_id=f'rupset_{fault_system_lt.short_name}',
                 )
             )
@@ -145,3 +94,45 @@ def test_composite_serialisation():
     assert new_composite.rates.columns.all() == composite.rates.columns.all()
     assert new_composite.rates.shape == composite.rates.shape
     assert new_composite.rates['Rupture Index'].all() == composite.rates['Rupture Index'].all()
+
+
+# class TestThreeFaultSystems(object):
+
+#     @pytest.mark.skip('REVIEW')
+#     def test_composite_rates_shape(self, three_branch_solutions):
+#         solutions = []
+
+#         v1_0_0 = nm.get_model_version('NSHM_1.0.0')
+#         slt = v1_0_0.source_logic_tree()
+#         print(slt.fault_system_branches[0])
+
+#         # for idx in [1, 2]:
+#         #     slt.fault_system_branches[idx].branches = deepcopy(slt.fault_system_branches[0].branches)
+
+#         # for fault_system_lt in slt.fault_system_branches:
+#         #     if fault_system_lt.short_name in ['CRU', 'PUY', 'HIK']:
+#         #         print('extending', fault_system_lt.short_name)
+#         #         solutions.extend(
+#         #             puy_branch_solutions(
+#         #                 fault_system_lt,
+#         #                 archive=ARCHIVES[fault_system_lt.short_name],
+#         #                 rupt_set_id=f'rupset_{fault_system_lt.short_name}',
+#         #             )
+#         #         )
+
+#         solutions = three_branch_solutions
+
+#         # print('solutions', len(solutions))
+#         # print(solutions)
+#         # print()
+
+#         composite = FaultSystemSolution.from_branch_solutions(solutions)
+#         # print( composite.aggregate_rates.info() )
+#         # print( composite.aggregate_rates.head() )
+#         # print( composite.aggregate_rates.tail() )
+
+#         # assert composite.rates.shape == (3101 + 15800 + 23675, RATE_COLUMNS_A)
+#         assert composite.rates.shape == (4211, RATE_COLUMNS_A)
+#         assert composite.composite_rates.shape == (3 * 4211, COMPOSITE_RATE_COLUMNS)
+#         # assert composite.composite_rates.shape == (127728, COMPOSITE_RATE_COLUMNS)
+
