+++ test/test_dataframe_serialisation.py
@@ -4,11 +4,14 @@ import os
 import pathlib
 import unittest
 import zipfile
+import tempfile
+import solvis
+import pytest
 
 import pandas as pd
 
-
 class TestRates(unittest.TestCase):
+
     def test_read_rates_csv(self):
         folder = pathlib.PurePath(os.path.realpath(__file__)).parent
         filename = pathlib.PurePath(folder, "fixtures/ModularAlpineVernonInversionSolution.zip")
@@ -33,3 +36,74 @@ class TestRates(unittest.TestCase):
         rr = rates_df2
         out1 = rr[rr['Annual Rate'] > rate].index
         assert out0.all() == out1.all()
+
+class TestSerialisation(object):
+    def test_write_to_archive_compatible(self, crustal_fixture, archives):
+
+        folder = tempfile.TemporaryDirectory()
+        # folder = pathlib.PurePath(os.path.realpath(__file__)).parent
+        new_path = pathlib.Path(folder.name, 'test_compatible_archive.zip')
+
+        # ref_solution = next(branch_solutions(fslt, archive=archives['CRU']))
+
+        fixture_folder = pathlib.PurePath(os.path.realpath(__file__)).parent / "fixtures"
+        ref_solution = pathlib.PurePath(fixture_folder, archives['CRU'])
+
+        # assert not new_path.exists()
+        # write the file
+        crustal_fixture.to_archive(str(new_path), ref_solution, compat=True)
+        assert new_path.exists()
+
+    def test_write_read_archive_compatible(self, crustal_fixture, archives):
+
+        folder = tempfile.TemporaryDirectory()
+        # folder = pathlib.PurePath(os.path.realpath(__file__)).parent
+        new_path = pathlib.Path(folder.name, 'test_compatible_archive.zip')
+
+        fixture_folder = pathlib.PurePath(os.path.realpath(__file__)).parent / "fixtures"
+        ref_solution = pathlib.PurePath(fixture_folder, archives['CRU'])
+
+        crustal_fixture.to_archive(str(new_path), ref_solution, compat=True)
+        read_sol = solvis.FaultSystemSolution.from_archive(new_path)
+
+        print(read_sol.rates)
+        print(crustal_fixture.rates)
+        assert read_sol.rates.columns.all() == crustal_fixture.rates.columns.all()
+        assert read_sol.rates.shape == crustal_fixture.rates.shape
+
+    def test_write_read_archive_compatible_composite_rates(self, crustal_fixture, archives):
+
+        folder = tempfile.TemporaryDirectory()
+        # folder = pathlib.PurePath(os.path.realpath(__file__)).parent
+        new_path = pathlib.Path(folder.name, 'test_compatible_archive.zip')
+
+        fixture_folder = pathlib.PurePath(os.path.realpath(__file__)).parent / "fixtures"
+        ref_solution = pathlib.PurePath(fixture_folder, archives['CRU'])
+
+        crustal_fixture.to_archive(str(new_path), ref_solution, compat=True)
+        read_sol = solvis.FaultSystemSolution.from_archive(new_path)
+
+        print(read_sol.composite_rates.info())
+        print(read_sol.composite_rates.columns)
+        print(read_sol.composite_rates)
+        assert read_sol.composite_rates.columns.all() == crustal_fixture.composite_rates.columns.all()
+        assert read_sol.composite_rates.shape == crustal_fixture.composite_rates.shape
+
+    def test_write_read_archive_incompatible(self, crustal_fixture, archives):
+
+        folder = tempfile.TemporaryDirectory()
+        # folder = pathlib.PurePath(os.path.realpath(__file__)).parent
+        new_path = pathlib.Path(folder.name, 'test_incompatible_archive.zip')
+
+        fixture_folder = pathlib.PurePath(os.path.realpath(__file__)).parent / "fixtures"
+        ref_solution = pathlib.PurePath(fixture_folder, archives['CRU'])
+
+        crustal_fixture.to_archive(str(new_path), ref_solution, compat=False)
+        read_sol = solvis.FaultSystemSolution.from_archive(new_path)
+
+        print(read_sol.rates)
+        print(crustal_fixture.rates)
+        assert read_sol.rates.columns.all() == crustal_fixture.rates.columns.all()
+        # NO the composite solutions have different rate structure
+        # assert read_sol.rates.shape == crustal_fixture.rates.shape
+        assert read_sol.rates['Rupture Index'].all() == crustal_fixture.rates['Rupture Index'].all()
